Implement insert command (#10)
Resolves #3.

Adds Collection implementation with insert support
Applies some fixes to the Database implementation

Reviewed-on: https://git.frisiapp.com/slendium/ocd-mariadb/pulls/10

+++ src/Database/DataType.php
@@ -17,7 +17,7 @@ enum DataType : string {
 
 	case Int = 'int';
 
-	case Decimal = 'decimal';
+	case Double = 'double';
 
 	case VarChar = 'varchar';
 
@@ -28,7 +28,7 @@ enum DataType : string {
 	public static function fromStorageClass(StorageClass $storageClass): self {
 		return match($storageClass) {
 			StorageClass::String => self::VarChar,
-			StorageClass::Float => self::Decimal,
+			StorageClass::Float => self::Double,
 			StorageClass::Int => self::Int,
 			StorageClass::Bool => self::TinyInt
 		};
